requirements.txt
Pillow==8.4.0
numpy==1.21.3
tensorflow==2.7.0
requests==2.26.0